fix: resolve ruff issues and bot recommendations
